pyproject.toml
[tool.poetry]
name = "drf-hw2-0"
version = "0.1.0"
description = ""
authors = ["DimaPopov9991 <popov9991@gmail.com>"]
readme = "README.md"
packages = [{include = "drf_hw2"}]

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"
django = "^5.0.7"
django-dotenv = "^1.4.2"
django-filter = "^24.3"



[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
